langchain-ask-csv-main/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
astor==0.8.1
async-timeout==4.0.3
attrs==25.3.0
blinker==1.9.0
boto3==1.37.13
botocore==1.37.13
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
duckdb==1.2.1
exceptiongroup==1.2.2
Faker==19.13.0
fonttools==4.56.0
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.39.0
google-genai==1.11.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.45
langchain-experimental==0.3.4
langchain-text-splitters==0.3.6
langsmith==0.3.13
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.9.4
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.30.0
numpy==2.0.2
openai==1.66.3
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pandasai==2.4.2
pillow==10.4.0
propcache==0.3.0
protobuf==5.29.3
py4j==0.10.9.7
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.9.1
pyparsing==3.2.1
pyspark==3.5.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.23.1
rsa==4.9.1
s3transfer==0.11.4
scipy==1.13.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.39
sqlglot==25.34.1
sqlglotrs==0.3.0
streamlit==1.43.2
streamlit-lottie==0.0.5
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
websockets==15.0.1
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
